requirements.txt
black==23.9.1
python-decouple==3.8
feedparser==6.0.10
discord.py==2.3.2
markdownify==0.11.6
